[e.cash] Fix X overflow scroll on homepage mobile
Summary:
I noticed on mobile was getting some x scrollbars coming from these two components
due to their x position animations. Adding some overflow hidden rules to prevent it

Also removing the initialX stuff that was creating the alternating side entrances. I was finding it too busy, especially on mobile, so simplifying to just all come in from the right, but open to keeping it if there are objections

Test Plan:
preview master and you should see a x scroll bar on mobile screens
This branch should not have it

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D18524