[Cashtab] Use sk directly in the Collections component instead of getting it from wif
Summary:
Cashtab used `wif` for private keys before the ecash-lib implementation. After ecash-lib, it uses hex. The wallet storage already keeps the format used by ecash-lib. So, there is no need to convert the wif to get this sk.

This is an artifact from before the wallet migration, when we only had wif stored, but we needed sk to build this specific kind of tx using ecash-agora.

Test Plan: `npm test`, we have existing tests that buy and cancel NFT offers, so this confirms maintenance of expected behavior

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18539