[Cashtab] Autoclose the webapp tx modal after 2s
Summary:
User feedback is that they do not like the extra click. I think though it is important to show that the tx worked.

So, continue showing the success modal. But also show a progress bar that auto-closes the modal after 2s.

Will see how this works. If there is still user concern, could make the close time a setting param, or could have a setting where users could disable the confirmation notification.

Test Plan: `npm test`, I tested this locally with the extension and the webapp, it is deployed at cashtab.io. However it is difficult to test without actually opening this modal from a webapp.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18552