[Cashtab] When opened for a param-populated tx, use AsOrdered chronik server instead of ClosestFirst
Summary:
For the use case of specific tx, we do not care about the "fastest" chronik server. We are only using chronik to send this one specific tx.

Higher impact to load the modal as fast as possible, and to use a consistent server (for, e.g., players making lots of consecutive txs).

Also testing to see if this fixes the "repeat txid" issue from rapid fire txs, or if that requires further optimizations. Regardless, we should still do this for modals, as racing chronik servers is not worth it for a single-action modal.

Test Plan: `npm test`, open an extension tx with the modal, check the logs, note no latency ordering, `npm start`, we do see latency ordering, can do the same by manually putting a '?' in the url

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18553