[ecash-lib] Modify GENESIS payment action to optionally accept group token id
Summary:
In adding support for minting NFTs to ecash-wallet, we will need a way for a user to specify this in their Action. Minting an NFT is a GENESIS tx of a SLP_TOKEN_TYPE_NFT1_CHILD. Per spec, the genesis tx must include a utxo of the group token at input index 0. The inclusion of this is what implies the group of this NFT.

So, the user must specify the group, so that ecash-wallet knows what utxo to select.

In Cashtab we only support minting from the group page, so we just have this info. But ecash-wallet will need this info from the user.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18557