[Cashtab] Patch spinner issue for first time users with no wallet
Summary:
The initial utxo sync has been working to solve its intended problem, that of wallets opening and receiving user input before utxo set is synced. However the case of new users with no wallet was never unsetting this flag.

Make sure we unset it.

Test Plan:
`npm start`, in an incognito window go to localhost:3000, note that the spinner is cleared

compare to current prod, go to cashtab.com in an incognito window and spinner hangs indefinitely

Reviewers: #bitcoin_abc, johnkuney

Reviewed By: #bitcoin_abc, johnkuney

Differential Revision: https://reviews.bitcoinabc.org/D18566