[avalanche] Fix the getavalancheinfo status when the local proof is avalanche invalidated
Summary:
This is another edge case that was not properly refecting the proof status: if there is a better conflicting proof on the network, the getavalancheinfo will properly return this status. But if this is missed by the opterator, and the conflicting proof disappears, and the node restarts (lots of if but it happened for real), the local proof might still be rejected by avalanche as the stakers remember the invalidation status. This leaves the RPC with no status information and a proof that won't verify as valid.

Fortunately because the node polls for this proof, we have the invalidation information. This diff makes it available to the RPC status so it shows something useful instead of "pending verification".

Test Plan:
  ./test/functional/test_runner.py abc_rpc_getavalancheinfo

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18593