[Cashtab] For webapp txs, make the modal more clearly wallet connect and not just copy address
Summary:
Just like with the extension, the user is sharing their address with an app. We are not just "Copying" the address. So, change the terminology. Add a reject button so the user can clearly opt out.

- We add a reject button so the user can clearly opt out of sharing address
- We change "Copy" to "Connect" which more accurately describes the work flow, we are ultimately sharing this address with an app

I looked into getting the requester but in the webapp there is not a good way to do this, as it could be spoofed. So imo best to just not say where the request is from.

Test Plan:
`npm test`, deploy to cashtab.io and test with webapp

check screenshot:
{F18124291}

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18595