[Cashtab] Use the min fee as the default for all users
Summary:
Now that we have strongly typed libs and all fee calculations are exact, there is no reason for fees to be a user setting. It should just be the min.

However, it is useful to keep it available as a setting, since this lets us mock different fees for different tests, which lets us show that we are not breaking past functionality at least in this diff. Also in the future, we may want to expose a fee setting.

Gamification and "VIP" features is a good concept but needs some more thought. The min fee thing isn't a good prize. It was rolled out this way when we were still unsure about the fee calc but that is no longer the case, all users should have min fees.

This diff

- Implements min fees for all Cashtab users
- Removes the VIP setting
- Introduces satsPerKb as a hidden setting. For now, not accessible to users, but can be used to mock different fee rates in tests (allows us to keep existing mocks in existing tests)

Test Plan:
`npm test`

this diff is deployed to cashtab.io

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18603