[e.cash] Improve powering payments section
Summary:
The powering payments section on the homepage desktop view was a bit breakable and buggy. Mainly moving the cursor around quickly could get it stuck in not the correct state. Trying to animate in all the details text in the center was causing these issues.

Proposing moving the text to overlay each card when hovered instead. This simplifies a lot, and removes some of the performance heavy animations, and is arguably better design since the information is visually attached to each card.

Shorten a couple of the texts to fit better.

Also fixed the vertical scroll on mobile screens if you click one of the icons. Before it would scroll the text to the top of the page, now it just stays where it is and scrolls horizontally into view.

Test Plan: preview and check out the section on homepage. Try to break it, note any performance issues

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D18608