[Cashtab] Remove multi-path support from tx history parsing and websocket subscriptions
Summary:
Depends on D18613

Deprecating support from Cashtab's multi-path wallets as part of the effort to use `ecash-wallet` in Cashtab.

For now, it's worth keeping `getHashes` around as a helper function for preparing test mocks. It's okay to keep support for a wallet with multiple hashes in tx history parsing, as we would expect this from an HD wallet.

But we stop using the function anywhere in the actual app. Stop subscribing to alt-path address websockets, stop parsing incoming tx notifications that could have been sent to alt path addresses.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18645