[e.cash] Improve what we do section images
Summary:
On the homepage, in the second section (+what we do}. The images have a little blue tag with a word over them. The tag was part of the image but this was causing some crunchiness on the text, and some mis-alignment between them.

Opting to instead create the tag with css so they can be vertically aligned, and the text will be more clear.

before
{F18181817}
after
{F18181816}

Test Plan: preview and check

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D18649