[Cashtab] Improve form inputs for wallet app
Summary:
T3883

Cashtab text inputs are used for addresses, bip21 strings, and mnemonics. We really do not want "normal" input rules for this type of info.

We have used "type=email" as a bit of a hack in mnemonic fields in order to prevent auto capitalization. Modern browsers though have some more robust and specific ways of handling this. Implement these throughout Cashtab text input components.

Goals
1. Prevent auto-capitalization and auto-completion in mnemonic input forms
2. Prevent the same in Send address forms (esp there is an issue where, user has an address on his clipboard, goes to fill address field, sees something suggested and assumes it is clipboard; but really it is is autocomplete based on a common address)

Test Plan: `npm test`, type a mnemonic and it is not capitalized, click an address field and it does not suggest autocomplete

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18658