[ecash-script] Deprecate ecash-script
Summary:
The functionality of ecash-script has been replicated in ecash-lib. We know longer have any monorepo apps depending on ecash-script.

Additional parsing methods will be built out in ecash-lib. ecash-script is legacy, it is still a JS lib, it has not been supported, and the OP_RETURN parsing and writing methods in ecash-lib are more sophisticated.

So, lose it. It has already been marked as deprecated on npm.

Test Plan: run cashtab and ecash-herald tests; CI failure expected for ecash-script with its removal

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18668