[Cashtab] Remove parsedTxHistory from wallet state
Summary:
Tx history is helpful information for the user. However, it is not necessary for any wallet actions. It should not be associated with the wallet state, especially now that we have paginated tx history which can only use the wallet state for page 1.

Remove the `parsedTxHistory` key from wallet state. Note we are no longer storing this key, so this change does not require a migration.

This allows us to implement more fine-grained handling of tx history and avalanche finality and achieve current performance without the hacky batchUpdates process. We could probably also lose our async queueing of websocket msgs, now that we are appropriately using Reacts setState with prev, but at the moment it works and is basically instant so will do that in another diff if it should be done.

This diff is additional prep for dropping in `ecash-wallet`. With wallet state `parsedTxHistory` no longer used by the app, we can use `ecash-wallet` instead of cashtab wallet state, once it has feature parity with Cashtab's custom methods.

Test Plan: `npm test`, `npm start`, send a bunch of txs, note they are instantly added to history, note if the page is full the last tx is removed from the page, note they all finalize at the same time with no issues. This diff is live at cashtab.io

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18669